Fixing stuff after review
