+++ go.mod
@@ -2,7 +2,4 @@ module go-yandex-metrics
 
 go 1.22.2
 
-require (
-	github.com/go-chi/chi/v5 v5.0.12
-	github.com/kelseyhightower/envconfig v1.4.0
-)
+require github.com/go-chi/chi/v5 v5.0.12
